refactor: 컨트롤러의 비즈니스로직 책임분리 1. List<List<Car>> 타입의 일급클래스 적용 2. 사용하지 않는 필드 제거 3. 정적 팩토리 매서드 패턴 적용 - 생성자 내부 비즈니스로직(createCarList()) 분리
