+++ src/main/java/racingcar/domain/Car.java
@@ -1,6 +1,6 @@
 package racingcar.domain;
 
-import racingcar.domain.strategyPattern.MoveStrategy;
+import racingcar.domain.strategy.MoveStrategy;
 
 public class Car {
     private static final int MIN_MOVE_COUNT = 4;
